Refactoring: class GeoDataPointmap::BreakLineFinishDefinitionCommand moved to independent files.
