+++ libs/geodata/pointmap/private/geodatapointmap_breaklinefinishdefinitioncommand.h
@@ -1,0 +1,21 @@
+#ifndef GEODATAPOINTMAP_BREAKLINEFINISHDEFINITIONCOMMAND_H
+#define GEODATAPOINTMAP_BREAKLINEFINISHDEFINITIONCOMMAND_H
+
+#include "../geodatapointmap.h"
+
+#include <QUndoCommand>
+
+class GeoDataPointmap::BreakLineFinishDefinitionCommand : public QUndoCommand
+{
+public:
+	BreakLineFinishDefinitionCommand(GeoDataPointmap* parent);
+
+	void redo() override;
+	void undo() override;
+
+private:
+	GeoDataPointmapBreakLine* m_breakLine;
+	GeoDataPointmap* m_pointMap;
+};
+
+#endif // GEODATAPOINTMAP_BREAKLINEFINISHDEFINITIONCOMMAND_H
